[SRP] Uma classe para gerar EPUB
